chore(app): move models to common
